services/authentication/requirements.txt
asgiref==3.10.0
cffi==2.0.0
cryptography==46.0.2
Django==5.2.7
django-cors-headers==4.3.1
django-tenants==3.9.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
jwcrypto==1.5.6
psycopg2-binary==2.9.10
pycparser==2.23
PyJWT==2.8.0
python-decouple==3.8
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
